requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
appdirs==1.4.4
argh==0.26.2
asgiref==3.2.9
astroid==2.9.3
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.4
boto3==1.20.29
botocore==1.23.29
bs4==0.0.1
cachetools==4.2.1
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
cryptography==36.0.1
cycler==0.10.0
Cython==0.29.23
dataframe-image==0.1.1
decorator==4.4.2
deepsnap==0.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.7
django-cors-headers==3.4.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.0.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gensim==4.1.0
getFamaFrenchFactors==0.0.5
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.32.0
# Editable install with no version control (gym==0.17.3)
-e d:\ovgu-4\project stowkoski\gym-master\gym-master
gym-stowkoski==0.0.1
h5py==2.10.0
huggingface-hub==0.2.1
idna==2.9
ijson==3.1.4
imbalanced-learn==0.7.0
imblearn==0.0
iniconfig==1.1.1
interchange==2021.0.4
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
isort==5.10.1
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.1
jupyterlab-widgets==1.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.7.1
littleutils==0.2.2
lxml==4.6.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mlxtend==0.17.2
monotonic==1.6
multidict==5.1.0
multitasking==0.0.10
mysql-connector-python==8.0.25
nbconvert==5.6.1
nbformat==5.0.5
neo4j==4.4.0
neo4j-driver==4.4.1
networkx==2.6.3
nltk==3.5
node2vec==0.4.3
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
ogb==1.3.2
opencv-python==4.4.0.46
openpyxl==3.0.6
opt-einsum==3.3.0
outdated==0.2.1
packaging==21.3
pandas==1.0.5
pandocfilters==1.4.2
pansi==2020.7.3
parso==0.7.0
patsy==0.5.2
pickleshare==0.7.5
Pillow==7.1.1
pkgtools==0.7.3
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.14.0
py==1.11.0
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydbgen==1.0.5
pyg==0.7.1
pyglet==1.5.0
Pygments==2.6.1
pylint==2.12.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.5
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.0
rdflib==6.0.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7
s3transfer==0.5.0
sacremoses==0.0.46
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.11.1
Send2Trash==1.5.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.15.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.0.1
sqlparse==0.3.1
stable-baselines3==0.10.0
statsmodels==0.13.1
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
textblob==0.15.3
textdistance==4.2.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
torch-bincount==0.1.1
torch-cluster==1.5.9
torch-geometric @ git+https://github.com/pyg-team/pytorch_geometric.git@57506bc2ce05e250eac5012994fc01296f60575f
torch-scatter==2.0.8
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.9.1
torchvision==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
transformers==4.13.0
typing_extensions==4.0.1
urllib3==1.25.9
virtualenv==20.4.1
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
xvfbwrapper==0.2.9
yacs==0.1.8
yarl==1.6.3
yfinance==0.1.67
